Add update for unsplit solving functionality.
This adds functionality for the flux solver to be able to modify the
input states for the flux solver, after the states are reconstructed but
before they are input to the flux method to solve the flux.

The unsplit solver implementation uses the the above addition to define
an implementation for solving multi-dimensional fluxes in a
dimensionally-unsplit manner. This is not yet tested for correctness.