Add a hacked together fix for ND corner loading.
This adds a very messy, but working, implementation of corner loading
for boundary data that will handle all cases. This is tricky because the
corners of a shared memory patch have different conditions. For example
in 2D, the following combinations are possible (X, Y):

- (boundary, boundary)
- (boundary, internal)
- (internal, boundary)
- (internal, internal)

Additionally, the case where the last block's boundary element's thread
indices are not multiples of the block size (for example, data of size
100 and block size 16) leads to threads in the middle of the thread
block needing to be treated as if they are actually boundary threads.
This implementation handles all these cases.