Add LevelSet, MultiIter fill and foreach.
This adds an initial LevelSet class implementation which can set the
initial values for a level set.

In addition, utility functions are added. fill() and for_each()
functions are added for multi dimensional iterators, as well as the
traits required for multi dimensional iterators to implement the
functionality.

NOTE: Code all needs to be cleaned.

NOTE: Unfixed 2d unsplit implementation is also present here.