Add Quadratic class and update Stencil interface
This commit adds a Quadratic class which can be used to store the
coefficients of a quadratic, solve the quadratic, and add and subtract
quadratics from each other.

It also updates the Stencil interface to allow the computation of
quadratic coefficients for a stencil for a given dimension.

A first order two point class which implements the Stencil inteface is
also added.