Add Box geometry class, and Box sdf eikonal test
This adds a Box geometry class, and refactores the Pos class and also
adds a Vec3 class which currently just aliases a 3D array.

It also adds  an Eikonal validation test for the box, and refactors the
original test into a Sphere Eikonal test.

This also adds various math functionality to the math module, and
refactors some of the array functionality to be slightly cleaner.